New updates (copies)

+++ teamods/debugging/make.sh
@@ -4,12 +4,18 @@ name=debug
 wrap=main_${name}
 repo=$HOME/vc/primer4/doc/src/chapters/debug
 
-rm -rf html *.html *.pdf
 #cp -r $repo/sphinx-rootdir/_build/html .
 #cp -r $repo/sphinx-rootdir-extended/_build/html html-githg
-cp $repo/$wrap.pdf $name.pdf
+
+if [ -f $repo/$wrap.pdf ]; then
+  rm -f $name.pdf
+  cp $repo/$wrap.pdf $name.pdf
+fi
 #cp $repo/${wrap}_4print.pdf ${name}_4print.pdf
+
+rm -rf html *.html
 cp $repo/${name}.html $repo/._${name}*.html .
 #cp $repo/${wrap}-solarized.html ${name}-solarized.html
+
 git add .
 git commit -am 'New updates (copies)'
